requirements.txt
alembic
anyio
aiofiles
bcrypt
cffi
click
cryptography
ecdsa
fastapi
h11
httptools
idna
jinja2       
passlib
psycopg2-binary
pyasn1
pycparser
pydantic
PyMySQL
python-dotenv
python-jose
python-multipart
PyYAML
rsa

sniffio
SQLAlchemy

typing_extensions
uvicorn
watchfiles
websocket
